+++ config/defaults.php
@@ -2,30 +2,30 @@
 /**
  * Genesis Starter Theme
  *
- * @package   SEOThemes\GenesisStarterTheme
+ * @package   SeoThemes\GenesisStarterTheme
  * @link      https://seothemes.com/genesis-starter-theme
  * @author    SEO Themes
  * @copyright Copyright © 2018 SEO Themes
  * @license   GPL-3.0-or-later
  */
 
-namespace SEOThemes\GenesisStarterTheme;
+namespace SeoThemes\GenesisStarterTheme;
 
-use D2\Core\AssetLoader;
-use D2\Core\Constants;
-use D2\Core\CustomColors;
-use D2\Core\GenesisLayout;
-use D2\Core\GenesisSettings;
-use D2\Core\GoogleFonts;
-use D2\Core\Hooks;
-use D2\Core\ImageSizes;
-use D2\Core\PluginActivation;
-use D2\Core\SimpleSocialIcons;
-use D2\Core\TextDomain;
-use D2\Core\ThemeSupport;
-use D2\Core\WidgetArea;
+use SeoThemes\Core\AssetLoader;
+use SeoThemes\Core\Constants;
+use SeoThemes\Core\CustomColors;
+use SeoThemes\Core\GenesisSettings;
+use SeoThemes\Core\GoogleFonts;
+use SeoThemes\Core\Hooks;
+use SeoThemes\Core\ImageSizes;
+use SeoThemes\Core\PageLayouts;
+use SeoThemes\Core\PluginActivation;
+use SeoThemes\Core\SimpleSocialIcons;
+use SeoThemes\Core\TextDomain;
+use SeoThemes\Core\ThemeSupport;
+use SeoThemes\Core\WidgetArea;
 
-$d2_assets = [
+$core_assets = [
 	AssetLoader::SCRIPTS => [
 		[
 			AssetLoader::HANDLE   => 'menus',
@@ -37,7 +37,7 @@ $d2_assets = [
 			AssetLoader::LOCALIZE => [
 				AssetLoader::LOCALIZEVAR  => 'genesis_responsive_menu',
 				AssetLoader::LOCALIZEDATA => [
-					'mainMenu'         => __( 'Menu', 'genesis-starter-theme' ),
+					'mainMenu'         => '<span class="hamburger"></span><span class="screen-reader-text">' . __( 'Menu', 'genesis-starter-theme' ) . '</span>',
 					'subMenu'          => __( 'Sub Menu', 'genesis-starter-theme' ),
 					'menuIconClass'    => null,
 					'subMenuIconClass' => null,
@@ -51,9 +51,17 @@ $d2_assets = [
 			],
 		],
 		[
-			AssetLoader::HANDLE  => 'script',
+			AssetLoader::HANDLE  => 'fitvids',
 			AssetLoader::URL     => AssetLoader::path( '/resources/js/script.js' ),
 			AssetLoader::DEPS    => [ 'jquery' ],
+			AssetLoader::VERSION => wp_get_theme()->get( 'Version' ),
+			AssetLoader::FOOTER  => true,
+			AssetLoader::ENQUEUE => true,
+		],
+		[
+			AssetLoader::HANDLE  => 'script',
+			AssetLoader::URL     => AssetLoader::path( '/resources/js/jquery.fitvids.js' ),
+			AssetLoader::DEPS    => [ 'fitvids' ],
 			AssetLoader::VERSION => wp_get_theme()->get( 'Version' ),
 			AssetLoader::FOOTER  => true,
 			AssetLoader::ENQUEUE => true,
@@ -61,7 +69,7 @@ $d2_assets = [
 	],
 ];
 
-$d2_constants = [
+$core_constants = [
 	Constants::DEFINE => [
 		'CHILD_THEME_NAME'    => wp_get_theme()->get( 'Name' ),
 		'CHILD_THEME_URL'     => wp_get_theme()->get( 'ThemeURI' ),
@@ -73,7 +81,7 @@ $d2_constants = [
 	],
 ];
 
-$d2_custom_colors = [
+$core_custom_colors = [
 	'background' => [
 		'default' => '#ffffff',
 		'output'  => [
@@ -89,11 +97,13 @@ $d2_custom_colors = [
 		],
 	],
 	'link'       => [
-		'default' => '#0073e5',
+		'default' => '#0077ee',
 		'output'  => [
 			[
 				'elements'   => [
 					'a',
+					'.site-title a:focus',
+					'.site-title a:hover',
 					'.entry-title a:focus',
 					'.entry-title a:hover',
 					'.genesis-nav-menu a:focus',
@@ -113,27 +123,24 @@ $d2_custom_colors = [
 		],
 	],
 	'accent'     => [
-		'default' => '#0073e5',
+		'default' => '#0077ee',
 		'output'  => [
 			[
 				'elements'   => [
 					'button:focus',
 					'button:hover',
-					'input[type="button"]:focus',
-					'input[type="button"]:hover',
-					'input[type="reset"]:focus',
-					'input[type="reset"]:hover',
-					'input[type="submit"]:focus',
-					'input[type="submit"]:hover',
-					'input[type="reset"]:focus',
-					'input[type="reset"]:hover',
-					'input[type="submit"]:focus',
-					'input[type="submit"]:hover',
+					'[type="button"]:focus',
+					'[type="button"]:hover',
+					'[type="reset"]:focus',
+					'[type="reset"]:hover',
+					'[type="submit"]:focus',
+					'[type="submit"]:hover',
+					'[type="reset"]:focus',
+					'[type="reset"]:hover',
+					'[type="submit"]:focus',
+					'[type="submit"]:hover',
 					'.button:focus',
 					'.button:hover',
-					'.genesis-nav-menu > .menu-highlight > a:hover',
-					'.genesis-nav-menu > .menu-highlight > a:focus',
-					'.genesis-nav-menu > .menu-highlight.current-menu-item > a',
 				],
 				'properties' => [
 					'background-color' => '%s',
@@ -143,26 +150,61 @@ $d2_custom_colors = [
 	],
 ];
 
-$d2_example = [
+$core_example = [
 	Example::SUB_CONFIG => [
 		Example::KEY => 'value',
 	],
 ];
 
-$d2_genesis_settings = [
+$core_genesis_settings = [
 	GenesisSettings::DEFAULTS => [
 		GenesisSettings::SITE_LAYOUT => 'full-width-content',
 	],
 ];
 
-$d2_google_fonts = [
+$core_google_fonts = [
 	GoogleFonts::ENQUEUE => [
 		'Source+Sans+Pro:400,600,700',
 	],
 ];
 
-$d2_hooks = [
+$core_hooks = [
 	Hooks::ADD    => [
+		[
+			Hooks::TAG      => 'wp_enqueue_scripts',
+			Hooks::CALLBACK => 'genesis_enqueue_main_stylesheet',
+			Hooks::PRIORITY => 99,
+		],
+		[
+			Hooks::TAG      => 'body_class',
+			Hooks::CALLBACK => function ( $classes ) {
+				if ( is_home() || is_search() || is_author() || is_date() || is_category() || is_tag() || is_page_template( 'page_blog.php' ) ) {
+					$classes[] = 'post-grid';
+				}
+
+				$classes[] = 'no-js';
+
+				return $classes;
+
+			},
+		],
+		[
+			Hooks::TAG      => 'genesis_before',
+			Hooks::CALLBACK => function () {
+				?>
+				<script>
+                    //<![CDATA[
+                    (function () {
+                        var c = document.body.classList;
+                        c.remove('no-js');
+                        c.add('js');
+                    })();
+                    //]]>
+				</script>
+				<?php
+			},
+			Hooks::PRIORITY => 1,
+		],
 		[
 			Hooks::TAG         => 'genesis_site_title',
 			Hooks::CALLBACK    => 'the_custom_logo',
@@ -210,6 +252,29 @@ $d2_hooks = [
 			},
 		],
 		[
+			Hooks::TAG      => 'genesis_structural_wrap-footer',
+			Hooks::CALLBACK => function ( $output, $original_output ) {
+				if ( 'open' == $original_output ) {
+					$output = '<div class="footer-credits">' . $output;
+				} elseif ( 'close' == $original_output ) {
+					$backtotop = '<a href="#" rel="nofollow" class="backtotop">Return to top</a>';
+					$output    = $backtotop . $output . $output;
+				}
+
+				return $output;
+			},
+			Hooks::PRIORITY => 10,
+			Hooks::ARGS     => 2,
+		],
+		[
+			Hooks::TAG      => 'genesis_before',
+			Hooks::CALLBACK => function () {
+				if ( 'center-content' === genesis_site_layout() ) {
+					add_filter( 'genesis_site_layout', '__genesis_return_full_width_content' );
+				}
+			}
+		],
+		[
 			Hooks::TAG      => 'admin_init',
 			Hooks::CALLBACK => function () {
 				add_editor_style( 'editor.css' );
@@ -230,6 +295,10 @@ $d2_hooks = [
 	],
 	Hooks::REMOVE => [
 		[
+			Hooks::TAG      => 'genesis_meta',
+			Hooks::CALLBACK => 'genesis_load_stylesheet',
+		],
+		[
 			Hooks::TAG      => 'genesis_after_header',
 			Hooks::CALLBACK => 'genesis_do_nav',
 		],
@@ -244,7 +313,7 @@ $d2_hooks = [
 	],
 ];
 
-$d2_image_sizes = [
+$core_image_sizes = [
 	ImageSizes::ADD => [
 		'featured' => [
 			'width'  => 620,
@@ -259,18 +328,25 @@ $d2_image_sizes = [
 	],
 ];
 
-$d2_layouts = [
-	GenesisLayout::UNREGISTER => [
-		// GenesisLayout::CONTENT_SIDEBAR,
-		// GenesisLayout::SIDEBAR_CONTENT,
-		// GenesisLayout::FULL_WIDTH_CONTENT,
-		GenesisLayout::CONTENT_SIDEBAR_SIDEBAR,
-		GenesisLayout::SIDEBAR_SIDEBAR_CONTENT,
-		GenesisLayout::SIDEBAR_CONTENT_SIDEBAR,
-	]
+$core_layouts = [
+	PageLayouts::REGISTER   => [
+		[
+			'id'    => 'center-content',
+			'label' => __( 'Center Content', 'genesis-starter-theme' ),
+			'img'   => get_stylesheet_directory_uri() . '/resources/img/center-content.gif',
+		]
+	],
+	PageLayouts::UNREGISTER => [
+		// PageLayouts::CONTENT_SIDEBAR,
+		// PageLayouts::SIDEBAR_CONTENT,
+		// PageLayouts::FULL_WIDTH_CONTENT,
+		PageLayouts::CONTENT_SIDEBAR_SIDEBAR,
+		PageLayouts::SIDEBAR_SIDEBAR_CONTENT,
+		PageLayouts::SIDEBAR_CONTENT_SIDEBAR,
+	],
 ];
 
-$d2_plugins = [
+$core_plugins = [
 	PluginActivation::REGISTER => [
 		[
 			PluginActivation::NAME     => 'Genesis Widget Column Classes',
@@ -295,22 +371,30 @@ $d2_plugins = [
 	],
 ];
 
-$d2_simple_social_icons = [
+if ( class_exists( 'WooCommerce' ) ) {
+	$core_plugins[ PluginActivation::REGISTER ][] = [
+		PluginActivation::NAME     => 'Genesis Connect for WooCommerce',
+		PluginActivation::SLUG     => 'genesis-connect-woocommerce',
+		PluginActivation::REQUIRED => false,
+	];
+}
+
+$core_simple_social_icons = [
 	SimpleSocialIcons::DEFAULTS => [
 		SimpleSocialIcons::NEW_WINDOW => 1,
 		SimpleSocialIcons::SIZE       => 40,
 	],
 ];
 
-$d2_textdomain = [
+$core_textdomain = [
 	TextDomain::DOMAIN => 'genesis-starter-theme',
 ];
 
-$d2_theme_support = [
+$core_theme_support = [
 	ThemeSupport::ADD => [
-		'align-wide',
-		'automatic-feed-links',
-		'custom-logo'              => [
+		'align-wide'                  => null,
+		'automatic-feed-links'        => null,
+		'custom-logo'                 => [
 			'height'      => 100,
 			'width'       => 300,
 			'flex-height' => true,
@@ -320,7 +404,7 @@ $d2_theme_support = [
 				'.site-description',
 			],
 		],
-		'custom-header'            => [
+		'custom-header'               => [
 			'header-selector' => '.hero-section',
 			'default_image'   => get_stylesheet_directory_uri() . '/resources/img/hero.jpg',
 			'header-text'     => false,
@@ -331,7 +415,7 @@ $d2_theme_support = [
 			'uploads'         => true,
 			'video'           => true,
 		],
-		'genesis-accessibility'    => [
+		'genesis-accessibility'       => [
 			'404-page',
 			'drop-down-menu',
 			'headings',
@@ -340,22 +424,22 @@ $d2_theme_support = [
 			'skip-links',
 		],
 		'genesis-after-entry-widget-area',
-		'genesis-footer-widgets'   => 4,
-		'genesis-menus'            => [
+		'genesis-footer-widgets'      => 3,
+		'genesis-menus'               => [
 			'primary'   => __( 'Header Menu', 'genesis-starter-theme' ),
 			'secondary' => __( 'After Header Menu', 'genesis-starter-theme' ),
 		],
-		'genesis-responsive-viewport',
-		'genesis-structural-wraps' => [
+		'genesis-responsive-viewport' => null,
+		'genesis-structural-wraps'    => [
 			'header',
 			'menu-secondary',
 			'footer-widgets',
 			'footer',
 		],
-		'gutenberg'                => [
+		'gutenberg'                   => [
 			'wide-images' => true,
 		],
-		'html5'                    => [
+		'html5'                       => [
 			'caption',
 			'comment-form',
 			'comment-list',
@@ -363,15 +447,15 @@ $d2_theme_support = [
 			'search-form',
 		],
 		'post-thumbnails',
-		'woocommerce',
-		'wc-product-gallery-zoom',
-		'wc-product-gallery-lightbox',
-		'wc-product-gallery-slider',
-		'wp-block-styles',
+		'woocommerce'                 => null,
+		'wc-product-gallery-zoom'     => null,
+		'wc-product-gallery-lightbox' => null,
+		'wc-product-gallery-slider'   => null,
+		'wp-block-styles'             => null,
 	],
 ];
 
-$d2_widget_areas = [
+$core_widget_areas = [
 	WidgetArea::UNREGISTER => [
 		WidgetArea::HEADER_RIGHT,
 		WidgetArea::SIDEBAR_ALT,
@@ -379,18 +463,18 @@ $d2_widget_areas = [
 ];
 
 return [
-	AssetLoader::class       => $d2_assets,
-	Constants::class         => $d2_constants,
-	CustomColors::class      => $d2_custom_colors,
-	Example::class           => $d2_example,
-	GenesisLayout::class     => $d2_layouts,
-	GenesisSettings::class   => $d2_genesis_settings,
-	GoogleFonts::class       => $d2_google_fonts,
-	Hooks::class             => $d2_hooks,
-	ImageSizes::class        => $d2_image_sizes,
-	PluginActivation::class  => $d2_plugins,
-	SimpleSocialIcons::class => $d2_simple_social_icons,
-	TextDomain::class        => $d2_textdomain,
-	ThemeSupport::class      => $d2_theme_support,
-	WidgetArea::class        => $d2_widget_areas,
+	AssetLoader::class       => $core_assets,
+	Constants::class         => $core_constants,
+	CustomColors::class      => $core_custom_colors,
+	Example::class           => $core_example,
+	GenesisSettings::class   => $core_genesis_settings,
+	GoogleFonts::class       => $core_google_fonts,
+	Hooks::class             => $core_hooks,
+	ImageSizes::class        => $core_image_sizes,
+	PageLayouts::class       => $core_layouts,
+	PluginActivation::class  => $core_plugins,
+	SimpleSocialIcons::class => $core_simple_social_icons,
+	TextDomain::class        => $core_textdomain,
+	ThemeSupport::class      => $core_theme_support,
+	WidgetArea::class        => $core_widget_areas,
 ];